Create ezdfstree.py
This is a python 3.3 exercise to recursively descend the filesystem from the current working directory. It is tested on android and on windows. I have not determined all of the environment variable settings for proper windows operation. It will handle fery large file systems with not too much memory because a partitioned sort is built into it.